requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
arcade==3.2.0
    # via wizard_vs_irs (pyproject.toml)
attrs==25.3.0
    # via pytiled-parser
cffi==1.17.1
    # via pymunk
pillow==11.0.0
    # via arcade
pycparser==2.22
    # via cffi
pyglet==2.1.6
    # via arcade
pymunk==6.9.0
    # via arcade
pypresence==4.3.0
    # via wizard_vs_irs (pyproject.toml)
pytiled-parser==2.2.9
    # via arcade
typing-extensions==4.14.1
    # via pytiled-parser
